Implemented bookCollection module
